// src/models/UpdateSemesterDetails.ts
export interface UpdateSemesterDetails {
    commission:         number;
    yearMoment:         string;
    startDate:          Date;
    classesAmount:      number;
    minimumAttendance:  number;
}

export interface UpdateSemesterDetailsSnakeCase {
    commission:         number;
    year_moment:        string;
    start_date:         Date;
    classes_amount:     number;
    minimum_attendance: number;
}